캠페인 메인 화면 및 상세 페이지 조회 기능 구현

+++ src/main/java/com/keypoint/keypointtravel/campaign/entity/CampaignBudget.java
@@ -30,10 +30,10 @@ public class CampaignBudget {
     private String category;
 
     @Column(nullable = false)
-    private Long amount;
+    private float amount;
 
     @Builder
-    public CampaignBudget(Campaign campaign, CurrencyType currency, String category, Long amount) {
+    public CampaignBudget(Campaign campaign, CurrencyType currency, String category, float amount) {
         this.campaign = campaign;
         this.currency = currency;
         this.category = category;
